Implement views for direct unsubscribing
The design is that the user clicks the unsubscribe link
to arrive at the landing page.
Here they select one out of two options.
The first option directly deletes the associated
Subscriber object and sends a confirmation email to
the subscriber.
The second option sends a standard token email
to the Subscriber so they can manage their subscriptions.